# swagger.yaml
swagger: "2.0"
info:
  description: "Software Engineering Project API"
  version: "1.0.0"
  title: "Shfuggiente SE 2"
host: "shfuggientese2.herokuapp.com"
basePath: "/api/v1"
schemes:
- "https"
consumes:
- application/json
produces:
- application/json
paths:
...
  /users:
    get:
      tags:
        - "Users"
      summary: Get a list of the users
      responses:
        200:
          description: 'successful response'
          examples:
            'application/json':
              - id: 1
                name: 'user1'
              - id: 2
                name: 'user2'
    post:
      summary: "Create a user"
      parameters:
      - in: "body"
        name: "User informations"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "User created successfully"
        400:
          description: "Error creating user"
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        type: integer
        minimum: 1
    get:
      summary: "Get user informations"
      responses:
        200:
          description: "successful response"
          schema:
            $ref: "#/definitions/User"
    put:
      summary: "Edit user"
      parameters:
      - in: "body"
        name: "User"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "User modified successfully"
    delete:
      summary: "Delete user"
      responses:
        200:
          description: 'Deleted successfully'
...
definitions:
...
  User:
    type: "object"
    required:
    - "name"
    - "mail"
    - "role"
    properties:
      id:
        type: "integer"
        minimum: 1
        format: "int64"
      name:
        type: "string"
        example: "Name 1"
      surname:
        type: "string"
        example: "Surname 1"
      role:
        type: "string"
        example: "Teacher"
      mail:
        type: "string"
        example: "mail@provider.net"
      groups:
        type: "array"
        example: [34,65,12,6]
...